Exposes additional Manifest structs
This PR exposes additional structs to allow consumers to better interact
with Manifests created by this crate. The additional structs are
`chunk::Chunk` as well as `chunk::manifest::FileChunkInfo`.

Additionally, removed the restrictions on exact version matching on
dependencies. Kept to all major version, but allowed minor upgrades to
be adopted.